requirements.txt
fastapi
uvicorn
pillow
pytesseract
opencv-python-headless
numpy